Let date functions have their own file
